requirements.txt
google-api-core
google-api-python-client
google-auth
google-cloud-core
google-cloud-storage
google-cloud-bigquery
pandas
Pillow
psycopg2-binary
PyYAML
SQLAlchemy
numpy
matplotlib
exdir